pipeline/requirements.txt
google-cloud-bigquery==3.20.1
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.49.0
aiohttp==3.9.5
Pillow==10.3.0
opencv-python-headless==4.9.0.80
torch==2.2.2
transformers==4.40.1
requests==2.31.0
